De-archangel: rename GradingClientModule/SandalphonClientModule; move Git to judgels.git
- GradingClientModule -> GradingModule
- SandalphonClientModule -> ResourceModule (judgels.resource)
- Move Git/LocalGit/GitCommit from judgels.sandalphon to judgels.git
- Move problem statement HTML resources from judgels/sandalphon to judgels/resource
