+++ judgels-backends/judgels-server-app/src/main/java/judgels/lesson/version/LessonVersionStore.java
@@ -10,8 +10,8 @@ import judgels.persistence.LessonDao;
 import judgels.persistence.LessonModel;
 import judgels.problem.base.ProblemFs;
 import judgels.problem.base.ProblemGit;
-import judgels.sandalphon.Git;
-import judgels.sandalphon.GitCommit;
+import judgels.git.Git;
+import judgels.git.GitCommit;
 
 public class LessonVersionStore extends BaseLessonStore {
     private final Git lessonGit;
