+++ judgels-backends/judgels-server-app/src/main/java/judgels/michael/lesson/version/LessonVersionResource.java
@@ -27,7 +27,7 @@ import judgels.michael.resource.ListVersionHistoryView;
 import judgels.michael.resource.RebaseVersionLocalChangesView;
 import judgels.michael.resource.ViewVersionLocalChangesView;
 import judgels.michael.template.HtmlTemplate;
-import judgels.sandalphon.GitCommit;
+import judgels.git.GitCommit;
 
 @Path("/lessons/{lessonId}/versions")
 public class LessonVersionResource extends BaseLessonResource {
